se puso el svg en un archivo independiente
