+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/card/index.js
@@ -1,3 +1,4 @@
-export * from "./card";
-
-export * from "./cardBody";
+export * from "./card.js";
+export * from "./cardHeader.js";
+export * from "./cardBody.js";
+export * from "../data/likes.js";
